% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveRedisTuning.R
\name{ArchiveRedisTuning}
\alias{ArchiveRedisTuning}
\title{Logging object for objective function evaluations}
\description{
Logging object for objective function evaluations

Logging object for objective function evaluations
}
\section{Super class}{
\code{\link[bbotk:Archive]{bbotk::Archive}} -> \code{ArchiveRedisTuning}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{(\link[redux:redis_config]{redux::redis_config})\cr
Configuration for connecting to Redis.}

\item{\code{n_workers}}{(\link{integer})\cr
Number of workers.}

\item{\code{instance_id}}{(\link{character})\cr
Unique identifier of the instance.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{connector}}{(\link[redux:hiredis]{redux::hiredis})\cr
Redis connector.}

\item{\code{all_keys}}{(\link{character})\cr
All keys of the archive.}

\item{\code{data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Contains all performed \link{Objective} function calls.}

\item{\code{terminated}}{(\link{logical})\cr
Whether the tuning process has been terminated.}

\item{\code{free_workers}}{(\link{integer})\cr
Number of free workers.}

\item{\code{n_evals}}{(\link{integer})\cr
Number of performed evaluations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArchiveRedisTuning-new}{\code{ArchiveRedisTuning$new()}}
\item \href{#method-ArchiveRedisTuning-write_xdt_xss}{\code{ArchiveRedisTuning$write_xdt_xss()}}
\item \href{#method-ArchiveRedisTuning-pop_xs}{\code{ArchiveRedisTuning$pop_xs()}}
\item \href{#method-ArchiveRedisTuning-write_ys}{\code{ArchiveRedisTuning$write_ys()}}
\item \href{#method-ArchiveRedisTuning-sync_data}{\code{ArchiveRedisTuning$sync_data()}}
\item \href{#method-ArchiveRedisTuning-best}{\code{ArchiveRedisTuning$best()}}
\item \href{#method-ArchiveRedisTuning-reset}{\code{ArchiveRedisTuning$reset()}}
\item \href{#method-ArchiveRedisTuning-clone}{\code{ArchiveRedisTuning$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="add_evals"><a href='../../bbotk/html/Archive.html#method-Archive-add_evals'><code>bbotk::Archive$add_evals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="clear"><a href='../../bbotk/html/Archive.html#method-Archive-clear'><code>bbotk::Archive$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="format"><a href='../../bbotk/html/Archive.html#method-Archive-format'><code>bbotk::Archive$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="nds_selection"><a href='../../bbotk/html/Archive.html#method-Archive-nds_selection'><code>bbotk::Archive$nds_selection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="print"><a href='../../bbotk/html/Archive.html#method-Archive-print'><code>bbotk::Archive$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$new(
  search_space,
  codomain,
  check_values = TRUE,
  config = redux::redis_config(),
  n_workers = future::nbrOfFreeWorkers(),
  instance_id = uuid::UUIDgenerate()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Hyperparameter search space. If \code{NULL} (default), the search space is
constructed from the \link{TuneToken} of the learner's parameter set
(learner$param_set).}

\item{\code{codomain}}{(\link[bbotk:Codomain]{bbotk::Codomain})\cr
Specifies codomain of objective function i.e. a set of performance measures.
Internally created from provided \link[mlr3:Measure]{mlr3::Measure}s.}

\item{\code{check_values}}{(\code{logical(1)})\cr
ignored.}

\item{\code{config}}{(\link[redux:redis_config]{redux::redis_config})\cr
Configuration for connecting to Redis.}

\item{\code{n_workers}}{(\code{integer(1)})\cr
Number of workers.}

\item{\code{instance_id}}{(\code{character(1)})\cr
Unique identifier of the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-write_xdt_xss"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-write_xdt_xss}{}}}
\subsection{Method \code{write_xdt_xss()}}{
Write \code{xdt} and \code{xss} to redis hashes and add keys to eval queue.
Both \code{xdt} and \code{xss} are serialized before being written to redis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$write_xdt_xss(xdt, xss)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{(\link[data.table:data.table]{data.table::data.table})\cr
Set of untransformed points and meta information.}

\item{\code{xss}}{(\link{list})\cr
List of transformed points.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-pop_xs"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-pop_xs}{}}}
\subsection{Method \code{pop_xs()}}{
Pop xs from eval queue and deserialize it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$pop_xs(timeout = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{(\link{integer})\cr
Timeout in seconds.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-write_ys"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-write_ys}{}}}
\subsection{Method \code{write_ys()}}{
Write ys to redis hash and add key to result queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$write_ys(key, ys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{(\code{character(1)})\cr
Key of the point.}

\item{\code{ys}}{(\code{list})\cr
List of results and meta information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-sync_data"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-sync_data}{}}}
\subsection{Method \code{sync_data()}}{
Sync data from redis to local data table.
Pops all keys from result queue and reads \code{"xdt"}, \code{"xss"} and \code{"ys"} from redis hashes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$sync_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-best"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-best}{}}}
\subsection{Method \code{best()}}{
Returns the best scoring evaluation(s).
For single-crit optimization, the solution that minimizes / maximizes the objective function.
For multi-crit optimization, the Pareto set / front.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$best(n_select = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select. Ignored for multi-crit optimization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-reset"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-reset}{}}}
\subsection{Method \code{reset()}}{
Resets the archive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRedisTuning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRedisTuning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRedisTuning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
