% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_search_space.R
\name{as_search_space}
\alias{as_search_space}
\alias{as_search_space.Learner}
\alias{as_search_space.ParamSet}
\alias{as_search_space.TuningSpace}
\title{Convert to a Search Space}
\usage{
as_search_space(x, ...)

\method{as_search_space}{Learner}(x, ...)

\method{as_search_space}{ParamSet}(x, ...)

\method{as_search_space}{TuningSpace}(x, ...)
}
\value{
\link[paradox:ParamSet]{paradox::ParamSet}.
}
\description{
Convert object to a search space.
}
