% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tuner.R
\name{as_tuner}
\alias{as_tuner}
\alias{as_tuner.Tuner}
\alias{as_tuners}
\alias{as_tuners.default}
\alias{as_tuners.list}
\title{Convert to a Tuner}
\usage{
as_tuner(x, ...)

\method{as_tuner}{Tuner}(x, clone = FALSE, ...)

as_tuners(x, ...)

\method{as_tuners}{default}(x, ...)

\method{as_tuners}{list}(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
Whether to clone the object.}
}
\description{
Convert object to a \link{Tuner} or a list of \link{Tuner}.
}
