% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackBatchTuning.R
\name{assert_batch_tuning_callback}
\alias{assert_batch_tuning_callback}
\alias{assert_batch_tuning_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_batch_tuning_callback(callback, null_ok = FALSE)

assert_batch_tuning_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link[bbotk:CallbackBatch]{bbotk::CallbackBatch} | \link{CallbackBatchTuning}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link[bbotk:CallbackBatch]{bbotk::CallbackBatch} | list of \link{CallbackBatchTuning}).}
}
\value{
\link[bbotk:CallbackBatch]{bbotk::CallbackBatch} | \link{CallbackBatchTuning} | List of them.
}
\description{
Assertions for \link{CallbackBatchTuning} class.
\link[bbotk:CallbackBatch]{bbotk::CallbackBatch} objects are accepted as well since they work in batch tuning too.
}
