% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calculate_pareto_front}
\alias{calculate_pareto_front}
\title{Return pareto front}
\usage{
calculate_pareto_front(points, maximize = TRUE)
}
\arguments{
\item{points}{Numeric matrix with each column corresponding to a point}

\item{maximize}{Bool vector of length one or nrow of points}
}
\description{
calculate pareto front of a matrix
also works for >2D pareto front
}
\examples{
\dontrun{
data = rbind(c(1, 0.25, 0.75, 0), c(0, 0.25, 0.75, 1))
calculate_pareto_front(data)
data = rbind(c(1,1,0,1), c(1,1,0,0), c(1,0,1,0), c(1,1,1,1))
calculate_pareto_front(data, maximize = FALSE)
}
}
