% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackTuning.R
\name{callback_tuning}
\alias{callback_tuning}
\title{Create Tuning Callback}
\usage{
callback_tuning(
  id,
  label = NA_character_,
  man = NA_character_,
  on_optimization_begin = NULL,
  on_optimizer_before_eval = NULL,
  on_eval_after_design = NULL,
  on_eval_after_benchmark = NULL,
  on_eval_before_archive = NULL,
  on_optimizer_after_eval = NULL,
  on_result = NULL,
  on_optimization_end = NULL,
  fields = list()
)
}
\arguments{
\item{on_optimization_begin}{(\verb{function()})\cr
Stage called at the beginning of the optimization.
Called in \code{Optimizer$optimize()}.
The functions must have two arguments named \code{callback} and \code{context}.}

\item{on_optimizer_before_eval}{(\verb{function()})\cr
Stage called after the optimizer proposes points.
Called in \code{OptimInstance$eval_batch()}.
The functions must have two arguments named \code{callback} and \code{context}.}

\item{on_eval_after_design}{(\verb{function()})\cr
Stage called after design is created.
Called in \code{ObjectiveTuning$eval_many()}.}

\item{on_eval_after_benchmark}{(\verb{function()})\cr
Stage called after hyperparameter configurations are evaluated.
Called in \code{ObjectiveTuning$eval_many()}.}

\item{on_eval_before_archive}{(\verb{function()})\cr
Stage called before performance values are written to the archive.
Called in \code{ObjectiveTuning$eval_many()}.}

\item{on_optimizer_after_eval}{(\verb{function()})\cr
Stage called after points are evaluated.
Called in \code{OptimInstance$eval_batch()}.
The functions must have two arguments named \code{callback} and \code{context}.}

\item{on_result}{(\verb{function()})\cr
Stage called after result are written.
Called in \code{OptimInstance$assign_result()}.
The functions must have two arguments named \code{callback} and \code{context}.}

\item{on_optimization_end}{(\verb{function()})\cr
Stage called at the end of the optimization.
Called in \code{Optimizer$optimize()}.
The functions must have two arguments named \code{callback} and \code{context}.}

\item{fields}{(list of \code{any})\cr
List of additional fields.}
}
\description{
Function to create a \link{CallbackTuning}.
}
\examples{
# write archive to disk
callback_tuning("mlr3tuning.backup",
  on_optimization_end = function(callback, context) {
    saveRDS(context$instance$archive, "archive.rds")
  }
)
}
