% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitnessFunction.R
\name{FitnessFunction}
\alias{FitnessFunction}
\title{FitnessFunction}
\description{
FitnessFunction.
}
\section{Usage}{
\preformatted{ff = FitnessFunction(id)
# public members
# public methods
# active bindings
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The id of the FitnessFunction.
\item \code{learner} (\code{Learner}):
The Learner that we want to evaluate.
\item \code{resampling} (\code{Resampling}):
The Resampling method that is used to obtain the y value.
\item \code{measure} (\code{Measure}):
Optional, can override the Measure in the Task
\item \code{param_set} (\link[paradox:ParamSet]{paradox::ParamSet}):
Parameter Set.
\item \code{terminator} (\code{Terminator}):
A \link{Terminator} which controls the termination of the tuning.
\item \code{tune_control} (\code{list()}):
See \code{\link[=tune_control]{tune_control()}}.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{FitnessFunction}.

\code{$eval(x)} (\code{numeric(length(self$measures))}) evaluates the parameter setting \code{x} (\code{list}) for the given learner and resampling.

\code{$eval_vectorized(xs)} (\code{matrix(length(xs), length(self$measures))}) performs resampling for multiple parameter settings \code{xs} (list of lists).
}

\concept{FitnessFunction}
\keyword{internal}
