% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{inner_tuning_results}
\alias{inner_tuning_results}
\title{Retrieve Inner Tuning Results}
\usage{
inner_tuning_results(resample_result)
}
\arguments{
\item{resample_result}{(\link[mlr3:ResampleResult]{mlr3::ResampleResult})\cr
Must contain an \link{AutoTuner}.}
}
\value{
\link[data.table:data.table]{data.table::data.table}.
}
\description{
Retrieves innner tuning results of nested resampling. The function iterates
over the \link{AutoTuner}s stored in the \link[mlr3:ResampleResult]{mlr3::ResampleResult} and binds the
tuning results to a \link[data.table:data.table]{data.table::data.table}.
}
\examples{
task = tsk("iris")
search_space = ParamSet$new(
  params = list(ParamDbl$new("cp", lower = 0.001, upper = 0.1))
)

at = AutoTuner$new(
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("evals", n_evals = 5),
  tuner = tnr("grid_search"),
  search_space = search_space,
  store_tuning_instance = TRUE)

resampling_outer = rsmp("cv", folds = 2)
rr = resample(task, at, resampling_outer, store_models = TRUE)

inner_tuning_results(rr)
}
