% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoTuner.R
\name{marshal_model.auto_tuner_model}
\alias{marshal_model.auto_tuner_model}
\title{Marshal AutoTuner}
\usage{
\method{marshal_model}{auto_tuner_model}(model, ...)
}
\arguments{
\item{model}{(model of \code{\link{AutoTuner}})\cr
The model to be marshaled.}
}
\description{
Modifies the model's \code{learner} in-place.
}
