% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3tuning-package}
\alias{mlr3tuning}
\alias{mlr3tuning-package}
\title{mlr3tuning: Tuning for mlr3}
\description{
Implements methods for hyperparameter tuning in mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3tuning.mlr-org.com}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3tuning/issues}
}

}
\author{
\strong{Maintainer}: Jakob Richter \email{jakob1richter@gmail.com}

Authors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.net}
  \item Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)
}

}
