% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3tuning.async_default_configuration}
\alias{mlr3tuning.async_default_configuration}
\title{Default Configuration Callback}
\description{
These \link{CallbackAsyncTuning} and \link{CallbackBatchTuning} evaluate the default hyperparameter values of a learner.
}
