% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3tuning.async_freeze_archive}
\alias{mlr3tuning.async_freeze_archive}
\title{Freeze Archive Callback}
\description{
This \link{CallbackAsync} freezes the \link{ArchiveAsync} to \link{ArchiveAsyncFrozen} after the optimization has finished.
}
\examples{
clbk("bbotk.async_freeze_archive")
}
