% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3tuning.early_stopping}
\alias{mlr3tuning.early_stopping}
\title{Early Stopping Callback}
\description{
When tuning an XGBoost learner, early stopping can be used to find the optimal number of trees.
}
\details{
This \link{Callback} extracts the maximum number of trees found during resampling.
The maximum number of trees are added to the \link{ArchiveTuning} in the column \code{"max_nrounds"}.
}
\examples{
clbk("mlr3tuning.early_stopping")
}
