% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr3tuning_assertions}
\alias{mlr3tuning_assertions}
\alias{assert_tuner}
\alias{assert_tuners}
\alias{assert_tuner_async}
\alias{assert_tuner_batch}
\alias{assert_tuning_instance}
\alias{assert_tuning_instance_async}
\alias{assert_tuning_instance_batch}
\title{Assertion for mlr3tuning objects}
\usage{
assert_tuner(tuner)

assert_tuners(tuners)

assert_tuner_async(tuner)

assert_tuner_batch(tuner)

assert_tuning_instance(inst)

assert_tuning_instance_async(inst)

assert_tuning_instance_batch(inst)
}
\arguments{
\item{tuner}{(\link{TunerBatch}).}

\item{tuners}{(list of \link{Tuner}).}

\item{inst}{(\link{TuningInstanceBatchSingleCrit} | \link{TuningInstanceBatchMultiCrit}).}
}
\description{
Most assertion functions ensure the right class attribute, and optionally additional properties.
}
\keyword{internal}
