% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerAsyncGridSearch.R
\name{mlr_tuners_async_grid_search}
\alias{mlr_tuners_async_grid_search}
\alias{TunerAsyncGridSearch}
\title{Hyperparameter Tuning with Asynchronous Grid Search}
\description{
Subclass for asynchronous grid search tuning.
}
\section{Dictionary}{

This \link{Tuner} can be instantiated with the associated sugar function \code{\link[=tnr]{tnr()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tnr("async_grid_search")
}\if{html}{\out{</div>}}
}

\section{Parameters}{


\describe{
\item{\code{batch_size}}{\code{integer(1)}\cr
Maximum number of points to try in a batch.}
}

}

\seealso{
Other TunerAsync:
\code{\link{mlr_tuners_async_design_points}},
\code{\link{mlr_tuners_async_random_search}}
}
\concept{TunerAsync}
\section{Super classes}{
\code{\link[mlr3tuning:Tuner]{Tuner}} -> \code{\link[mlr3tuning:TunerAsync]{TunerAsync}} -> \code{\link[mlr3tuning:TunerAsyncFromOptimizerAsync]{TunerAsyncFromOptimizerAsync}} -> \code{TunerAsyncGridSearch}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TunerAsyncGridSearch-initialize}{\code{TunerAsyncGridSearch$new()}}
    \item \href{#method-TunerAsyncGridSearch-clone}{\code{TunerAsyncGridSearch$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="format"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-format'><code>Tuner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="help"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-help'><code>Tuner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="print"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-print'><code>Tuner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TunerAsyncFromOptimizerAsync" data-id="optimize"><a href='../../mlr3tuning/html/TunerAsyncFromOptimizerAsync.html#method-TunerAsyncFromOptimizerAsync-optimize'><code>TunerAsyncFromOptimizerAsync$optimize()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerAsyncGridSearch-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TunerAsyncGridSearch-initialize}{}}}
\subsection{\code{TunerAsyncGridSearch$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerAsyncGridSearch$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerAsyncGridSearch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TunerAsyncGridSearch-clone}{}}}
\subsection{\code{TunerAsyncGridSearch$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerAsyncGridSearch$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
