% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerBatchCmaes.R
\name{mlr_tuners_cmaes}
\alias{mlr_tuners_cmaes}
\alias{TunerBatchCmaes}
\title{Hyperparameter Tuning with Covariance Matrix Adaptation Evolution Strategy}
\source{
Hansen N (2016).
\dQuote{The CMA Evolution Strategy: A Tutorial.}
1604.00772.
}
\description{
Subclass for Covariance Matrix Adaptation Evolution Strategy (CMA-ES).
Calls \code{\link[libcmaesr:cmaes]{libcmaesr::cmaes()}} from package \CRANpkg{libcmaesr}.
}
\section{Dictionary}{

This \link{Tuner} can be instantiated with the associated sugar function \code{\link[=tnr]{tnr()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tnr("cmaes")
}\if{html}{\out{</div>}}
}

\section{Progress Bars}{

\verb{$optimize()} supports progress bars via the package \CRANpkg{progressr}
combined with a \link[bbotk:Terminator]{bbotk::Terminator}. Simply wrap the function in
\code{progressr::with_progress()} to enable them. We recommend to use package
\CRANpkg{progress} as backend; enable with \code{progressr::handlers("progress")}.
}

\section{Logging}{

All \link{Tuner}s use a logger (as implemented in \CRANpkg{lgr}) from package
\CRANpkg{bbotk}.
Use \code{lgr::get_logger("bbotk")} to access and control the logger.
}

\section{Optimizer}{

This \link{Tuner} is based on \link[bbotk:OptimizerBatchCmaes]{bbotk::OptimizerBatchCmaes} which can be applied on any black box optimization problem.
See also the documentation of \href{https://bbotk.mlr-org.com/}{bbotk}.
}

\section{Parameters}{


\describe{
\item{\code{start_values}}{\code{character(1)}\cr
Create \code{"random"} start values or based on \code{"center"} of search space?
In the latter case, it is the center of the parameters before a trafo is applied.
If set to \code{"custom"}, the start values can be passed via the \code{start} parameter.}
\item{\code{start}}{\code{numeric()}\cr
Custom start values. Only applicable if \code{start_values} parameter is set to \code{"custom"}.}
\item{\code{seed}}{\code{integer(1)}\cr
Seed of the random number generator of \code{libcmaes}.
Unset by default, in which case the generator is seeded from R and the optimization is reproducible with
\code{\link[=set.seed]{set.seed()}}.}
}

All remaining parameters are passed to \code{\link[libcmaesr:cmaes_control]{libcmaesr::cmaes_control()}}, see there for their meaning.
Note that we have removed all control parameters which refer to the termination of the algorithm and where our
terminators allow to obtain the same behavior, i.e. \code{max_fevals}, \code{max_iter}, and \code{ftarget}.
The internal convergence criteria of the algorithm still apply, so the optimization can stop before the
\link[bbotk]{Terminator} is triggered.

}

\section{Batch evaluation}{


The optimizer evaluates a whole generation of \code{lambda} points in one batch.
The \link[bbotk]{Terminator} is only checked between generations, so the number of evaluations can exceed the budget of
\link[bbotk]{TerminatorEvals} by up to \code{lambda - 1} points.

}

\section{Resources}{

There are several sections about hyperparameter optimization in the \href{https://mlr3book.mlr-org.com}{mlr3book}.
\itemize{
\item Getting started with
\href{https://mlr3book.mlr-org.com/chapters/chapter4/hyperparameter_optimization.html}{hyperparameter optimization}.
\item An overview of all tuners can be found on our \href{https://mlr-org.com/tuners.html}{website}.
\item \href{https://mlr3book.mlr-org.com/chapters/chapter4/hyperparameter_optimization.html#sec-model-tuning}{Tune}
a support vector machine on the Sonar data set.
\item Learn about \href{https://mlr3book.mlr-org.com/chapters/chapter4/hyperparameter_optimization.html#sec-defining-search-spaces}{tuning spaces}.
\item Estimate the model performance with \href{https://mlr3book.mlr-org.com/chapters/chapter4/hyperparameter_optimization.html#sec-nested-resampling}{nested resampling}.
\item Learn about \href{https://mlr3book.mlr-org.com/chapters/chapter5/advanced_tuning_methods_and_black_box_optimization.html#sec-multi-metrics-tuning}{multi-objective optimization}.
\item Simultaneously optimize hyperparameters and use
\href{https://mlr3book.mlr-org.com/chapters/chapter15/predsets_valid_inttune.html}{early stopping} with XGBoost.
\item \href{https://mlr3book.mlr-org.com/chapters/chapter4/hyperparameter_optimization.html#sec-autotuner}{Automate}
the tuning.
}

The \href{https://mlr-org.com/gallery-all-optimization.html}{gallery} features a collection of case studies and demos
about optimization.
\itemize{
\item Learn more advanced methods with the \href{https://mlr-org.com/gallery/series/2021-03-09-practical-tuning-series-tune-a-support-vector-machine/}{Practical Tuning Series}.
\item Learn about \href{https://mlr-org.com/gallery/optimization/2023-01-16-hotstart/}{hotstarting} models.
\item Run the
\href{https://mlr-org.com/gallery/optimization/2023-01-31-default-configuration/}{default hyperparameter configuration}
of learners as a baseline.
\item Use the \href{https://mlr-org.com/gallery/series/2023-01-15-hyperband-xgboost/}{Hyperband} optimizer with different
budget parameters.
}

The \href{https://cheatsheets.mlr-org.com/mlr3tuning.pdf}{cheatsheet} summarizes
the most important functions of mlr3tuning.
}

\examples{
# example only runs if libcmaesr is available
if (mlr3misc::require_namespaces("libcmaesr", quietly = TRUE)) {
# Hyperparameter Optimization

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE),
  minsplit = to_tune(p_dbl(2, 128, trafo = as.integer)),
  minbucket = to_tune(p_dbl(1, 64, trafo = as.integer))
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("cmaes"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10)

# best performing hyperparameter configuration
instance$result

# all evaluated hyperparameter configuration
as.data.table(instance$archive)

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))
}
}
\seealso{
Other Tuner:
\code{\link{Tuner}},
\code{\link{mlr_tuners}},
\code{\link{mlr_tuners_design_points}},
\code{\link{mlr_tuners_gensa}},
\code{\link{mlr_tuners_grid_search}},
\code{\link{mlr_tuners_internal}},
\code{\link{mlr_tuners_irace}},
\code{\link{mlr_tuners_nloptr}},
\code{\link{mlr_tuners_random_search}}
}
\concept{Tuner}
\section{Super classes}{
\code{\link[mlr3tuning:Tuner]{Tuner}} -> \code{\link[mlr3tuning:TunerBatch]{TunerBatch}} -> \code{\link[mlr3tuning:TunerBatchFromOptimizerBatch]{TunerBatchFromOptimizerBatch}} -> \code{TunerBatchCmaes}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TunerBatchCmaes-initialize}{\code{TunerBatchCmaes$new()}}
    \item \href{#method-TunerBatchCmaes-clone}{\code{TunerBatchCmaes$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="format"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-format'><code>Tuner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="help"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-help'><code>Tuner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="print"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-print'><code>Tuner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TunerBatchFromOptimizerBatch" data-id="optimize"><a href='../../mlr3tuning/html/TunerBatchFromOptimizerBatch.html#method-TunerBatchFromOptimizerBatch-optimize'><code>TunerBatchFromOptimizerBatch$optimize()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerBatchCmaes-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TunerBatchCmaes-initialize}{}}}
\subsection{\code{TunerBatchCmaes$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerBatchCmaes$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerBatchCmaes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TunerBatchCmaes-clone}{}}}
\subsection{\code{TunerBatchCmaes$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerBatchCmaes$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
