% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoTuner.R
\name{set_validate.AutoTuner}
\alias{set_validate.AutoTuner}
\title{Configure Validation for AutoTuner}
\usage{
\method{set_validate}{AutoTuner}(learner, validate, validate_tuning, ...)
}
\arguments{
\item{learner}{(\code{\link{AutoTuner}})\cr
The autotuner for which to enable validation.}

\item{validate}{(\code{numeric(1)}, \code{"inner_valid"}, \code{"test"}, or \code{NULL})\cr
How to configure the validation of the \code{AutoTuner} itself (usually not enabled).}

\item{validate_tuning}{(\code{numeric(1)}, \code{"inner_valid"}, \code{"test"} or \code{NULL})\cr
How to set the validation during the tuning of the wrapped \code{\link{Learner}}.}
}
\description{
Configure Validation for AutoTuner
}
