% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorBase.R
\name{TerminatorBase}
\alias{TerminatorBase}
\title{TerminatorBase}
\description{
TerminatorBase.
}
\section{Usage}{
\preformatted{l = Terminator(id)
# public members
# public methods
# active bindings
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The name of the Terminator.
\item \code{settings} (\code{list}):
The settings for this Terminator.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorBase}.

\code{$update()} is called in each tuning iteration before the evaluation.

\code{$state} (\code{list}):
Custom state of the Terminator.
Individual for each subclass.
Gets updated with each call of \code{update()}.

\code{$terminated} (\code{logical(1)}):
Updated by each call of \code{update()}.
Is the termination criterion met?

\code{$message} a meaningfull chararacter string describing the state of the Terminator.
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorIterations}},
  \code{\link{TerminatorMultiplexer}}
}
\concept{Terminator}
\keyword{internal}
