% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvals.R
\docType{data}
\name{TerminatorEvals}
\alias{TerminatorEvals}
\title{Terminator that stops after a number of evals}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning depending on the number of evaluations.
An evaluation is defined by one resampling of a parameter value.
}
\section{Construction}{
\preformatted{t = TerminatorEvals$new(evals)
}
\itemize{
\item \code{evals} :: \code{integer(1)}\cr
Number of allowed evaluations.
Stored in \code{settings}.
}
}

\examples{
TerminatorEvals$new(3)
}
\seealso{
Other Terminator: \code{\link{TerminatorCombo}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
