% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvaluations.R
\name{TerminatorEvaluations}
\alias{TerminatorEvaluations}
\title{TerminatorEvaluations}
\description{
TerminatorEvaluations.
}
\section{Usage}{
\preformatted{t = TerminatorEvaluations$new(max_evaluations)
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{max_evaluations} [`integer(1)]:\cr
Maximum number of function evaluations.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorEvaluations}.
The interface is described in \link{Terminator}.
}

\examples{
t = TerminatorEvaluations$new(3)
}
\seealso{
Other Terminator: \code{\link{TerminatorIterations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{Terminator}}
}
\concept{Terminator}
