% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorIterations.R
\name{TerminatorIterations}
\alias{TerminatorIterations}
\title{TerminatorIterations Class}
\description{
Class to terminate the tuning depending on the number of iterations.
}
\section{Usage}{
\preformatted{# Constructor
t = TerminatorIterations$new(max_iterations)
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{max_iterations} (`integer(1)):
Maximum number of iterations.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorIterations}.
The interface is described in \link{Terminator}.
}

\examples{
t = TerminatorIterations$new(3)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{Terminator}}
}
\concept{Terminator}
