% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorModelTime.R
\docType{data}
\name{TerminatorModelTime}
\alias{TerminatorModelTime}
\title{Terminator that stops after a budget of model evaluation time is depleted}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a given model evaluation budget is exceeded.
The terminator measures the used time to train and predict all models contained
in the archive.
}
\section{Construction}{
\preformatted{t = TerminatorModelTime$new(secs)
}
\itemize{
\item \code{secs} :: \code{numeric(1)}\cr
Maximum allowed time, in seconds.
Stored in \code{settings}.
}
}

\examples{
TerminatorModelTime$new(3)
TerminatorModelTime$new(10 * 3600)
}
\seealso{
Other Terminator: \code{\link{TerminatorClockTime}},
  \code{\link{TerminatorCombo}},
  \code{\link{TerminatorEvals}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
