% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorMultiplexer.R
\name{TerminatorMultiplexer}
\alias{TerminatorMultiplexer}
\title{TerminatorMultiplexer}
\description{
TerminatorMultiplexer takes multiple Terminators and will lead to termination as soon as one Terminator signals a termination.
In the future more advanced termination rules can be implemented using this Multiplexer.
}
\section{Usage}{
\preformatted{t = TerminatorMultiplexer$new(terminators)
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{terminators} (\code{list}):
List of objects of class \link{Terminator}.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorMultiplexer}.
}

\examples{
t = TerminatorMultiplexer$new(list(
   TerminatorIterations$new(3),
   TerminatorEvaluations$new(10)
))
print(t)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorIterations}},
  \code{\link{Terminator}}
}
\concept{Terminator}
