% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorNone.R
\docType{data}
\name{TerminatorNone}
\alias{TerminatorNone}
\title{Terminator that never stops.}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Mainly useful for grid search, or maybe other tuners, where the stopping is inherently controlled by the tuner itself.
}
\section{Construction}{
\preformatted{t = TerminatorNone$new()
}
}

\seealso{
Other Terminator: \code{\link{TerminatorEvals}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
