% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerformance.R
\name{TerminatorPerformance}
\alias{TerminatorPerformance}
\title{TerminatorPerformance Class}
\description{
Class to terminate the tuning after some time. Note that the runtime is checked after each step
and therefore it could happen that the final runtime is longer than the specified one. Time is measured
for everything that happens between update_start and update_end.
}
\section{Usage}{
\preformatted{# Constructor
t = TerminatorPerformance$new(thresh, pe, all_reached = FALSE)
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{thresh} (named list):\cr
Thresholds that needs to be reached.
\item \code{pe} (\link{PerformanceEvaluator}):\cr
Performance evaluator used for the tuning. This is required to check whether the defined thresholds makes sense or not.
\item \code{all_reached} (logical(1)):\cr
Stop whether all thresholds are reached (default = FALSE).
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorRuntime}.

The interface is described in \link{Terminator}.
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorRuntime}}, \code{\link{Terminator}}
}
\concept{Terminator}
