% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerfReached.R
\docType{data}
\name{TerminatorPerfReached}
\alias{TerminatorPerfReached}
\title{Terminator with Performance Criterion}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a performance threshold has been hit.
}
\section{Construction}{
\preformatted{t = TerminatorPerfReached$new(thresh, pe)
}
\itemize{
\item \code{thresh} :: named \code{numeric()}\cr
Thresholds that need to be reached, named with measure ids.
Terminates if the performance exceeds (respective measure has to be maximized) or falls below (respective measure has to be minimized) all provided threshold simultaneously.
\item \code{pe} :: \link{PerformanceEvaluator}\cr
Performance evaluator used for the tuning.
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
