% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorRuntime.R
\name{TerminatorRuntime}
\alias{TerminatorRuntime}
\title{TerminatorRuntime Class}
\description{
Class to terminate the tuning after some time. Note that the runtime is checked after each step
and therefore it could happen that the final runtime is longer than the specified one.
}
\section{Usage}{
\preformatted{# Constructor
t = TerminatorRuntime$new(max_time, time_unit)
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{max_time} (integer(1)):
Maximal amount of time measures in \code{units}.
\item \code{units} (character(1)):
Unit used for measuring time. Possible choices are "secs", "mins", "hours", "days", and "weeks" that
are directly passed to \code{difftime()}.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorRuntime}.

The interface is described in \link{Terminator}.
}

\examples{
t = TerminatorRuntime$new(3, "mins")
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorIterations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{Terminator}}
}
\concept{Terminator}
