% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorStagnation.R
\docType{data}
\name{TerminatorStagnation}
\alias{TerminatorStagnation}
\alias{mlr_terminators_stagnation}
\title{Terminator that stops when a performance level has been reached}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after the performance stagnates, i.e. does not improve more than
\code{tol} over the last \code{iters} iterations.

TODO: NA handling.
}
\section{Construction}{
\preformatted{t = TerminatorStagnation$new(iters, tol = sqrt(.Machine$double.eps))
}
\itemize{
\item \code{iters} :: \code{integer(1)}\cr
Number of iterations to evaluate the performance improvement on.
\item \code{tol} :: \code{numeric(1)}\cr
Tolerance threshold. If the improvement is less than \code{tol}, tuning is stopped.
}
}

\seealso{
Other Terminator: \code{\link{TerminatorClockTime}},
  \code{\link{TerminatorCombo}},
  \code{\link{TerminatorEvals}},
  \code{\link{TerminatorModelTime}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{Terminator}}, \code{\link{mlr_terminators}},
  \code{\link{mlr_tuners}}
}
\concept{Terminator}
\keyword{datasets}
