% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorTime.R
\docType{data}
\name{TerminatorTime}
\alias{TerminatorTime}
\title{Terminator with Time Criterion}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a fixed date / time (as reported by \code{\link[=Sys.time]{Sys.time()}}.
}
\section{Construction}{
\preformatted{t = TerminatorTime$new(time)
}
\itemize{
\item \code{time} :: \code{POSIXct(1)}\cr
Terminator stops after this point in time.
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
time = Sys.time() + (60 * 60) # now + 1 hour
t = TerminatorTime$new(time)
print(t)

time = as.POSIXct("2030-01-01 00:00:00")
t = TerminatorTime$new(time)
print(t)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorPerformance}},
  \code{\link{TerminatorRuntime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
