% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{tune_auto}
\alias{tune_auto}
\title{Syntactic Sugar for Automatic Tuning}
\usage{
tune_auto(
  method,
  learner,
  resampling,
  measure,
  search_space = NULL,
  term_evals = NULL,
  term_time = NULL,
  ...
)
}
\description{
Sugar
}
