% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{tune_nested}
\alias{tune_nested}
\title{Syntactic Sugar for Nested Resampling}
\usage{
tune_nested(
  method,
  task,
  learner,
  inner_resampling,
  outer_resampling,
  measure,
  term_evals = NULL,
  term_time = NULL,
  search_space = NULL,
  ...
)
}
\description{
Sugar
}
