% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerAsyncFromOptimizerAsync.R
\name{TunerAsyncFromOptimizerAsync}
\alias{TunerAsyncFromOptimizerAsync}
\title{TunerAsyncFromOptimizerAsync}
\description{
Internally used to transform \link[bbotk:Optimizer]{bbotk::Optimizer} to \link{Tuner}.
}
\keyword{internal}
\section{Super classes}{
\code{\link[mlr3tuning:Tuner]{Tuner}} -> \code{\link[mlr3tuning:TunerAsync]{TunerAsync}} -> \code{TunerAsyncFromOptimizerAsync}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{param_set}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Set of control parameters.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TunerAsyncFromOptimizerAsync-initialize}{\code{TunerAsyncFromOptimizerAsync$new()}}
    \item \href{#method-TunerAsyncFromOptimizerAsync-optimize}{\code{TunerAsyncFromOptimizerAsync$optimize()}}
    \item \href{#method-TunerAsyncFromOptimizerAsync-clone}{\code{TunerAsyncFromOptimizerAsync$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="format"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-format'><code>Tuner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="help"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-help'><code>Tuner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="print"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-print'><code>Tuner$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerAsyncFromOptimizerAsync-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TunerAsyncFromOptimizerAsync-initialize}{}}}
\subsection{\code{TunerAsyncFromOptimizerAsync$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerAsyncFromOptimizerAsync$new(optimizer, man = NA_character_)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{optimizer}}{\link[bbotk:Optimizer]{bbotk::Optimizer}\cr
Optimizer that is called.}
      \item{\code{man}}{(\code{character(1)})\cr
String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.
The referenced help package can be opened via method \verb{$help()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerAsyncFromOptimizerAsync-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-TunerAsyncFromOptimizerAsync-optimize}{}}}
\subsection{\code{TunerAsyncFromOptimizerAsync$optimize()}}{
  Performs the tuning on a \link{TuningInstanceAsyncSingleCrit} /
\link{TuningInstanceAsyncMultiCrit} until termination. The single evaluations and
the final results will be written into the \link{ArchiveAsyncTuning} that
resides in the \link{TuningInstanceAsyncSingleCrit}/\link{TuningInstanceAsyncMultiCrit}.
The final result is returned.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerAsyncFromOptimizerAsync$optimize(inst)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{inst}}{(\link{TuningInstanceAsyncSingleCrit} | \link{TuningInstanceAsyncMultiCrit}).}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \link[data.table:data.table]{data.table::data.table}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerAsyncFromOptimizerAsync-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TunerAsyncFromOptimizerAsync-clone}{}}}
\subsection{\code{TunerAsyncFromOptimizerAsync$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TunerAsyncFromOptimizerAsync$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
