% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerBase.R
\name{TunerBase}
\alias{TunerBase}
\title{TunerBase}
\description{
TunerBase.
}
\section{Usage}{
\preformatted{l = TunerBase(id)
# public members
# public methods
# active bindings
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The id of the Tuner.
\item \code{settings} (\code{list}):
The settings for the Tuner.
\item \code{terminator} (\code{Terminator}).
All tuning problems optimized with this Tuner object will be terminated by this terminator.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TunerBase}.
}

\seealso{
Other Tuner: \code{\link{TunerRandomSearch}}
}
\concept{Tuner}
\keyword{internal}
