% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGenSA.R
\docType{data}
\name{TunerGenSA}
\alias{TunerGenSA}
\alias{mlr_tuners_gensa}
\title{TunerGenSA}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Subclass for generalized simulated annealing tuning with \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
}
\section{Construction}{
\preformatted{TunerGenSA$new()
tnr("gensa")
}
}

\section{Parameters}{

For the meaning of the control param see \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
For consistency, we have removed all GenSA control parameters which refer to the stopping of the algorithm and
where mlr3 terminators allow to obtain the same behavior.
\itemize{
\item \code{smooth} :: \code{logical(1)}\cr
\item \code{temperature} :: \code{numeric(1)}\cr
\item \code{acceptance.param} :: \code{numeric(1)}\cr
\item \code{verbose} :: \code{logical(1)}\cr
\item \code{trace.mat} :: \code{logical(1)}\cr
}
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: \code{\link{TunerDesignPoints}},
  \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
