% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGridSearch.R
\docType{data}
\name{TunerGridSearch}
\alias{TunerGridSearch}
\title{TunerGridSearch}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Tuner child class to conduct a grid search.
}
\section{Construction}{
\preformatted{tuner = TunerGridSearch$new(pe, terminator, resolution)
}

For arguments, see \link{Tuner}, and additionally:
\itemize{
\item \code{resolution} :: \code{integer(1)}\cr
Resolution of the grid.
If none is specified we will try to calculate the resolution form the Terminator.
}
}

\section{Fields}{

See \link{Tuner}.
}

\section{Methods}{

See \link{Tuner}.
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
