% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerNestedResampling.R
\name{TunerNestedResampling}
\alias{TunerNestedResampling}
\title{TunerNestedResampling}
\description{
Tuner child class to conduct grid search.
}
\section{Usage}{
\preformatted{tuner = TunerNestedResampling$new(inner_tuner, outer_resampling)
}

See \link{Tuner} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{inner_tuner} (\link{Tuner}):
Tuner that is executed on each iteration of the outer resampling.
\item \code{outer_resampling} (\link{Resampling}):
Resampling object that controls the outer tuning.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TunerNestedResampling}.
The interface is described in \link{Tuner}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("holdout")
measures = mlr3::mlr_measures$mget("mmce")
task$measures = measures
param_set = paradox::ParamSet$new(
  params = list(paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1)))

ff = FitnessFunction$new(task, learner, resampling, param_set)

terminator = TerminatorEvaluations$new(5)
rs = TunerRandomSearch$new(ff, terminator_eval)

outer = mlr3::mlr_resamplings$get("cv")

nested = TunerNestedResampling$new(rs, outer)
nested$tune()
nested$ff$bmr$data
}
\seealso{
Other Tuner: \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{internal}
