% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerRandomSearch.R
\docType{data}
\name{TunerRandomSearch}
\alias{TunerRandomSearch}
\title{TunerRandomSearch}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Tuner child class to conduct a random search.
}
\section{Construction}{
\preformatted{tuner = TunerRandomSearch$new(pe, terminator, batch_size = 100L)
}

For arguments, see \link{Tuner}, and additionally:
\itemize{
\item \code{batch_size} :: \code{integer(1)}\cr
Maximum number of configurations to try in a batch.
Each batch is possibly executed in parallel via \code{\link[mlr3:benchmark]{mlr3::benchmark()}}.
}
}

\section{Fields}{

See \link{Tuner}.
}

\section{Methods}{

See \link{Tuner}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv")
resampling$param_set$values$folds = 2
measures = mlr3::mlr_measures$mget("classif.ce")
param_set = paradox::ParamSet$new(
  params = list(
    paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1)
  )
)
pe = PerformanceEvaluator$new(task, learner, resampling, measures, param_set)

terminator = TerminatorEvaluations$new(10)
rs = TunerRandomSearch$new(pe, terminator)
rs$tune()$tune_result()
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerGridSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
