% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningInstanceSingleCrit.R
\name{wrapper_async}
\alias{wrapper_async}
\title{Wrapper}
\usage{
wrapper_async(objective, archive)
}
\arguments{
\item{objective}{(\link{ObjectiveTuning})\cr
Objective function.}

\item{archive}{(\link{ArchiveTuning})\cr
Archive of evaluated points.}
}
\description{
Wrapper function for asynchronous evaluation of the objective function.
}
