% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackAsyncTuning.R
\name{assert_async_tuning_callback}
\alias{assert_async_tuning_callback}
\alias{assert_async_tuning_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_async_tuning_callback(callback, null_ok = FALSE)

assert_async_tuning_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link{CallbackAsyncTuning}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{CallbackAsyncTuning}).}
}
\value{
[CallbackAsyncTuning | List of \link{CallbackAsyncTuning}s.
}
\description{
Assertions for \link{CallbackAsyncTuning} class.
}
