% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackBatchTuning.R
\name{assert_batch_tuning_callback}
\alias{assert_batch_tuning_callback}
\alias{assert_batch_tuning_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_batch_tuning_callback(callback, null_ok = FALSE)

assert_batch_tuning_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link{CallbackBatchTuning}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{CallbackBatchTuning}).}
}
\value{
[CallbackBatchTuning | List of \link{CallbackBatchTuning}s.
}
\description{
Assertions for \link{CallbackBatchTuning} class.
}
