% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoTuner.R
\name{AutoTuner}
\alias{AutoTuner}
\title{AutoTuner}
\description{
The \link{AutoTuner} conducts tuning and sets the learners parameter configuration to the best parameters obtained by the tuning.

Additionally, this class can be used to do nested resampling by passing an \link{AutoTuner} object to resample.
}
\section{Usage}{
\preformatted{# Construction
at = AutoTuner$new(learner, resampling, param_set, terminator,
  tuner, tuner_settings, ctrl = tune_control(), id = "autotuner")

# public fields
at$learner
}

See \link{Learner} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{learner} (\link{Learner}): \cr
Internal learner that is tuned and finally returned as trained learner with optimal parameter configuration.
\item \code{resampling} (\link{Resampling}): \cr
Resampling strategy for the tuning.
\item \code{param_set} (\link[paradox:ParamSet]{paradox::ParamSet}) \cr
Parameter set for the tuning.
\item \code{terminator} (\link{Terminator}) \cr
Terminator used to stop the tuning.
\item \code{tuner} (Tuner Class Generator) \cr
Uninitialized tuner factory, e.g. TunerGridSearch.
\item \code{tuner_settings} (named list) \cr
List with tuner settings (e.g. see \code{?TunerGridSearch})
}
}

\section{Details}{

\itemize{
\item With \code{at$learner} the raw learner with the best parameter configuration can be accessed.
}

The interface is described in \link{Learner}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("holdout")
measures = mlr3::mlr_measures$mget("classif.ce")
task$measures = measures
param_set = paradox::ParamSet$new(
  params = list(paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1)))

terminator = TerminatorEvaluations$new(5)

at = AutoTuner$new(learner, resampling, param_set, terminator, tuner = TunerGridSearch,
  tuner_settings = list(resolution = 10L))

at$train(task)
at$learner
}
\concept{Learner}
\keyword{internal}
