% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoTuner.R
\name{marshal_auto_tuner}
\alias{marshal_auto_tuner}
\alias{marshal_model.auto_tuner_model}
\alias{unmarshal_model.auto_tuner_model_marshaled}
\title{(Un-)Marshal AutoTuner Model}
\usage{
\method{marshal_model}{auto_tuner_model}(model, inplace = FALSE, ...)

\method{unmarshal_model}{auto_tuner_model_marshaled}(model, inplace = FALSE, ...)
}
\arguments{
\item{model}{(model of \code{\link{AutoTuner}})\cr
The model to be marshaled.}

\item{inplace}{(\code{logical(1)})\cr
Whether to marshal in-place.}

\item{...}{(any)\cr
Currently unused.}
}
\description{
(Un-)marshal the model of a \link{AutoTuner}.
}
\keyword{internal}
