% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3tuning-package}
\alias{mlr3tuning}
\alias{mlr3tuning-package}
\title{mlr3tuning: Tuning for 'mlr3'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Hyperparameter optimization package of the mlr3 ecosystem. It features highly configurable search spaces via the paradox package and finds optimal hyperparameter configurations for any mlr3 learner. mlr3tuning works with several optimization algorithms e.g. Random Search, Iterated Racing, Bayesian Optimization (in mlr3mbo) and Hyperband (in mlr3hyperband). Moreover, it can automatically optimize learners and estimate the performance of optimized models with nested resampling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3tuning.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3tuning}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3tuning/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Daniel Schalk \email{daniel.schalk@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0003-0950-1947}{ORCID})
}

}
