% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3tuning-package}
\alias{mlr3tuning}
\alias{mlr3tuning-package}
\title{mlr3tuning: Tuning for 'mlr3'}
\description{
Implements methods for hyperparameter tuning with
    'mlr3', e.g. Grid Search, Random Search, or Simulated Annealing.
    Various termination criteria can be set and combined.  The class
    'AutoTuner' provides a convenient way to perform nested resampling in
    combination with 'mlr3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3tuning.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3tuning}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3tuning/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

Authors:
\itemize{
  \item Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Daniel Schalk \email{daniel.schalk@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0003-0950-1947}{ORCID})
  \item Marc Becker \email{marcbecker@posteo.de}
}

}
