% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3tuning.async_freeze_archive}
\alias{mlr3tuning.async_freeze_archive}
\title{Freeze Archive Callback}
\description{
This \link{CallbackAsyncTuning} freezes the \link{ArchiveAsyncTuning} to \link{ArchiveAsyncTuningFrozen} after the optimization has finished.
}
\examples{
clbk("mlr3tuning.async_freeze_archive")
}
