% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3tuning.backup}
\alias{mlr3tuning.backup}
\title{Backup Benchmark Result Callback}
\description{
This \link{Callback} writes the \link{BenchmarkResult} after each batch to disk.
}
\examples{
clbk("mlr3tuning.backup", path = "backup.rds")
}
