% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3tuning.early_stopping}
\alias{mlr3tuning.early_stopping}
\title{Early Stopping Callback}
\description{
This callback extract the optimal number of trees from XGBoost models and

When training an XGBoost learner, early stopping can be used to find the optimal number of trees.
}
\examples{
cllb("mlr3tuning.early_stopping")
}
