% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_terminators.R
\docType{data}
\name{mlr_terminators}
\alias{mlr_terminators}
\title{Dictionary of Terminators}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{Terminator}.
Each terminator has an associated help page, see \code{mlr_terminators_[id]}.

This dictionary can get populated with additional terminators by add-on packages.

For a more convenient way to retrieve and construct terminator, see \code{\link[=term]{term()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
term("evals", n_evals = 10)
}
\seealso{
Sugar function: \code{\link[=term]{term()}}

Other Terminator: \code{\link{TerminatorClockTime}},
  \code{\link{TerminatorCombo}},
  \code{\link{TerminatorEvals}},
  \code{\link{TerminatorModelTime}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorStagnation}},
  \code{\link{Terminator}}, \code{\link{mlr_tuners}}
}
\concept{Terminator}
\keyword{datasets}
