% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tuners.R
\docType{data}
\name{mlr_tuners}
\alias{mlr_tuners}
\title{Dictionary of Tuners}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\usage{
mlr_tuners
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{Tuner}.
Each tuner has an associated help page, see \code{mlr_tuners_[id]}.

This dictionary can get populated with additional tuners by add-on packages.

For a more convenient way to retrieve and construct tuner, see
\code{\link[=tnr]{tnr()}}/\code{\link[=tnrs]{tnrs()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
mlr_tuners$get("grid_search")
tnr("random_search")
}
\seealso{
Sugar functions: \code{\link[=tnr]{tnr()}}, \code{\link[=tnrs]{tnrs()}}
}
\keyword{datasets}
