% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerIrace.R
\name{mlr_tuners_irace}
\alias{mlr_tuners_irace}
\alias{TunerIrace}
\title{TunerIrace}
\description{
Subclass for iterated racing tuning calling \code{\link[irace:irace]{irace::irace()}} from package \CRANpkg{irace}.
}
\section{Parameters}{

\itemize{
\item \code{show.irace.output} (\code{logical(1)})
\item \code{debugLevel} (\code{integer(1)})
\item \code{seed} (\code{integer(1)})
\item \code{postselection} (\code{numeric(1)})
\item \code{elitist} (\code{integer(1)})
\item \code{elitistLimit} (\code{integer(1)})
\item \code{nbIterations} (\code{integer(1)})
\item \code{nbExperimentsPerIteration} (\code{integer(1)})
\item \code{minNbSurvival} (\code{integer(1)})
\item \code{nbConfigurations} (\code{integer(1)})
\item \code{mu} (\code{integer(1)})
\item \code{softRestart} (\code{integer(1)})
\item \code{softRestartThreshold} (\code{numeric(1)})
\item \code{digits} (\code{integer(1)})
\item \code{testType} (\code{character(1)})
\item \code{firstTest} (\code{integer(1)})
\item \code{eachTest} (\code{integer(1)})
\item \code{confidence} (\code{numeric(1)})
\item \code{capping} (\code{integer(1)})
\item \code{cappingType} (\code{character(1)})
\item \code{boundType} (\code{character(1)})
\item \code{boundMax} (\code{numeric(1)})
\item \code{boundDigits} (\code{integer(1)})
\item \code{boundPar} (\code{numeric(1)})
\item \code{boundAsTimeout} (\code{numeric(1)})
}

\code{show.irace.output} suppresses the output from \code{\link[irace:irace]{irace::irace()}} and only prints output from this package.
For the meaning of all other parameters, see \code{\link[irace:defaultScenario]{irace::defaultScenario()}}.
Note that we have removed all control parameters which refer to the termination of the algorithm and
where our terminators allow to obtain the same behavior.
}

\section{Dictionary}{

This \link{Tuner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_tuners} or with the associated sugar function \code{\link[=tnr]{tnr()}}:\preformatted{TunerIrace$new()
mlr_tuners$get("irace")
tnr("irace")
}
}

\examples{
# see ?Tuner
}
\references{
Birattari, M., Stützle, T., Paquete, L., & Varrentrapp, K. (2002).
A Racing Algorithm for Configuring Metaheuristics.
In Gecco (Vol. 2).
}
\seealso{
Other Tuner: 
\code{\link{mlr_tuners_cmaes}},
\code{\link{mlr_tuners_design_points}},
\code{\link{mlr_tuners_gensa}},
\code{\link{mlr_tuners_grid_search}},
\code{\link{mlr_tuners_nloptr}},
\code{\link{mlr_tuners_random_search}}
}
\concept{Tuner}
\section{Super class}{
\code{\link[mlr3tuning:Tuner]{mlr3tuning::Tuner}} -> \code{TunerIrace}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TunerIrace$new()}}
\item \href{#method-clone}{\code{TunerIrace$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="format">}\href{../../mlr3tuning/html/Tuner.html#method-format}{\code{mlr3tuning::Tuner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="optimize">}\href{../../mlr3tuning/html/Tuner.html#method-optimize}{\code{mlr3tuning::Tuner$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="print">}\href{../../mlr3tuning/html/Tuner.html#method-print}{\code{mlr3tuning::Tuner$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerIrace$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerIrace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
