% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerNloptr.R
\name{mlr_tuners_nloptr}
\alias{mlr_tuners_nloptr}
\alias{TunerNloptr}
\title{TunerNloptr}
\description{
Subclass for general nonlinear optimization procedures \code{\link[nloptr:nloptr]{nloptr::nloptr()}} from package \CRANpkg{nloptr}.
}
\section{Parameters}{

\itemize{
\item \code{algorithm} (\code{character(1)}).
\item \code{xtol_rel} (\code{numeric(1)}).
\item \code{xtol_abs} (\code{numeric(1)}).
\item \code{ftol_rel} (\code{numeric(1)}).
\item \code{ftol_abs} (\code{numeric(1)}).
\item \code{eval_g_eq} (`function).
\item \code{eval_g_ineq} (\code{function}).
}

For the meaning of the control parameters, see \code{\link[nloptr:nloptr]{nloptr::nloptr()}} and the \href{https://nlopt.readthedocs.io/en/latest/}{nlopt documentation}.
Note that we have removed all control parameters which refer to the termination of the algorithm and
where our terminators allow to obtain the same behavior.
Note that no gradient information is passed to \code{\link[nloptr:nloptr]{nloptr::nloptr()}}, and thus in general \code{algorithm}
settings that require gradients do not work.
}

\section{Dictionary}{

This \link{Tuner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_tuners} or with the associated sugar function \code{\link[=tnr]{tnr()}}:\preformatted{mlr_tuners$get("nloptr")
tnr("nloptr")
}
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: 
\code{\link{Tuner}},
\code{\link{mlr_tuners_design_points}},
\code{\link{mlr_tuners_gensa}},
\code{\link{mlr_tuners_grid_search}},
\code{\link{mlr_tuners_random_search}},
\code{\link{mlr_tuners}}
}
\concept{Tuner}
\section{Super class}{
\code{\link[mlr3tuning:Tuner]{mlr3tuning::Tuner}} -> \code{TunerNloptr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TunerNloptr$new()}}
\item \href{#method-clone}{\code{TunerNloptr$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="format">}\href{../../mlr3tuning/html/Tuner.html#method-format}{\code{mlr3tuning::Tuner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="print">}\href{../../mlr3tuning/html/Tuner.html#method-print}{\code{mlr3tuning::Tuner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="tune">}\href{../../mlr3tuning/html/Tuner.html#method-tune}{\code{mlr3tuning::Tuner$tune()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerNloptr$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerNloptr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
