% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoTuner.R
\name{set_validate.AutoTuner}
\alias{set_validate.AutoTuner}
\title{Configure Validation for AutoTuner}
\usage{
\method{set_validate}{AutoTuner}(learner, validate, tune_validate, ...)
}
\arguments{
\item{learner}{(\code{\link{AutoTuner}})\cr
The autotuner for which to enable validation.}

\item{validate}{(\code{numeric(1)}, \code{"inner_valid"}, \code{"test"}, or \code{NULL})\cr
How to configure the validation of the \code{AutoTuner} itself (usually not enabled).}

\item{tune_validate}{(\code{numeric(1)}, \code{"inner_valid"}, \code{"test"} or \code{NULL})\cr
How to set the validation during the tuning of the wrapped \code{\link{Learner}}.}

\item{...}{(any)\cr
Passed when calling \code{set_validate()} on the wrapped leaerner.}
}
\description{
Configure validation for the auto-tuner and the wrapped learner.
In most cases, the validation data for the \code{\link{AutoTuner}} itself should be \code{NULL} and one only wants to
configure the validation data during the hyperparameter optimization via the parameter \code{tune_validate}.
}
\examples{
at = auto_tuner(
  tuner = tnr("random_search"),
  learner = lrn("classif.debug", early_stopping = TRUE, iter = to_tune(upper = 1000L, internal = TRUE), validate = 0.2),
  resampling = rsmp("holdout")
)
# use the test set of the tuning resampling for validation, don't do validation for the AutoTuner itself
set_validate(at,
  validate = NULL,
  tune_validate = "test"
)
}
