% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Terminator.R
\name{Terminator}
\alias{Terminator}
\title{Terminator Base Class}
\description{
Base class for termination criteria.
}
\section{Usage}{
\preformatted{t = Terminator$new(settings)
# public members
t$terminated
t$settings
t$state
# public methods
t$update_start(ff)
t$update_end(ff)
# active bindings
t$remaining
}
}

\section{Arguments}{

\itemize{
\item \code{settings} (\code{list}):
The settings for Terminator as named list.
}
}

\section{Details}{

\itemize{
\item \code{$new()} creates a new object of class \link{Terminator}.
\item \code{$update_start()} is called in each tuning iteration before the evaluation.
\item \code{$update_end()} is called in each tuning iteration after the evaluation.
\item \code{$state} (\code{list}):
Arbitrary state of the Terminator. Gets updated with each call of \code{update_start()} and \code{update_end()}.
\item \code{$terminated} (\code{logical(1)}):
Is the termination criterion met?
Updated by each call of \code{update_start()}/\code{update_end()}.
}
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorIterations}},
  \code{\link{TerminatorMultiplexer}}
}
\concept{Terminator}
