% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Terminator.R
\docType{data}
\name{Terminator}
\alias{Terminator}
\title{Abstract Terminator Class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Abstract \code{Terminator} class that implements the main functionality each terminator must have.
A terminator is an object that determines when to stop the tuning.
}
\section{Construction}{
\preformatted{t = Terminator$new(settings)
}
\itemize{
\item \code{settings} :: named \code{list()}\cr
Arbitrary settings required by the child class.
}
}

\section{Fields}{

\itemize{
\item \code{settings} :: named \code{list()}\cr
Settings passed during construction.
\item \code{terminated} :: \code{logical(1)}\cr
Is \code{TRUE} if the termination criterion if the child class is met.
\item \code{state} :: \code{list()}\cr
Arbitrary state, depending on the child class.
}
}

\section{Methods}{

\itemize{
\item \code{update_start(pe)}\cr
\link{PerformanceEvaluator} -> \code{self}\cr
Is called in each tuning iteration before the evaluation.
\item \code{update_end(pe)}\cr
\link{PerformanceEvaluator} -> \code{self}\cr
Is called in each tuning iteration after the evaluation.
}
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorPerformance}},
  \code{\link{TerminatorRuntime}}
}
\concept{Terminator}
\keyword{datasets}
