% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Terminator.R
\docType{data}
\name{Terminator}
\alias{Terminator}
\title{Abstract Terminator Class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Abstract \code{Terminator} class that implements the main functionality each terminator must have.
A terminator is an object that determines when to stop the tuning.

Termination of tuning works as follows:
\itemize{
\item Evaluations in a tuner a performed in batches of size \code{batch_size}.
\item After a batch-eval the Tuner is checked, if it is positive, we stop.
\item The tuning algorithm itself might decide not to produce any more points or even a smaller batch in its last evaluation.
}

Therefore the following note seems in order: While it is definitely possible to execute a fine-grained control for termination,
and for many tuners we can specify exactly when to stop, it might happen that a few too many or even a few too few evaluations are
performed, especially if your batchsize is large. So better check the size of the returned archive.
}
\section{Construction}{
\preformatted{t = Terminator$new(settings)
}
\itemize{
\item \code{settings} :: named \code{list()}\cr
Arbitrary settings required by the child class.
}
}

\section{Fields}{

\itemize{
\item \code{settings} :: named \code{list()}\cr
Settings passed during construction.
\item \code{terminated} :: \code{logical(1)}\cr
Is \code{TRUE} if the termination criterion if the child class is met.
}
}

\section{Methods}{

\itemize{
\item \code{eval_before(pe)}\cr
\link{PerfEval} -> \code{self}\cr
Is called in each tuning iteration before the evaluation.
Can/Should be implemented in a subclass.
\item \code{eval_after(pe)}\cr
\link{PerfEval} -> \code{self}\cr
Is called in each tuning iteration after the evaluation.
Can/Should be implemented in a subclass.
}
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}
}
\concept{Terminator}
\keyword{datasets}
