% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorClockTime.R
\docType{data}
\name{TerminatorClockTime}
\alias{TerminatorClockTime}
\alias{mlr_terminators_clock_time}
\title{Terminator that stops according to the clock time}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning either after the complete process took a number of seconds on the clock
or a fixed time point has been reached (as reported by \code{\link[=Sys.time]{Sys.time()}}).
}
\section{Construction}{
\preformatted{TerminatorClockTime$new(secs = NULL, stop_time = NULL)
term("clock_time")
}
\itemize{
\item \code{secs} :: \code{numeric(1)}\cr
Maximum allowed time, in seconds.
Mutually exclusive with argument \code{stop_time}.
Stored in the parameter set \code{$param_set}.
\item \code{stop_time} :: \code{POSIXct(1)}\cr
Terminator stops after this point in time.
Mutually exclusive with argument \code{secs}.
Stored in the parameter set \code{$param_set}.
}
}

\examples{
term("clock_time", secs = 1800)

stop_time = as.POSIXct("2030-01-01 00:00:00")
term("clock_time", stop_time = stop_time)
}
\seealso{
Other Terminator: \code{\link{TerminatorCombo}},
  \code{\link{TerminatorEvals}},
  \code{\link{TerminatorModelTime}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorStagnation}},
  \code{\link{Terminator}}, \code{\link{mlr_terminators}},
  \code{\link{mlr_tuners}}
}
\concept{Terminator}
\keyword{datasets}
