% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorCombo.R
\docType{data}
\name{TerminatorCombo}
\alias{TerminatorCombo}
\alias{mlr_terminators_combo}
\title{Combine Terminators}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
This class takes multiple \link{Terminator}s and terminates as soon as one or all of the included terminators are positive.
}
\section{Construction}{
\preformatted{t = TerminatorCombo$new(terminators, any = TRUE)
}
\itemize{
\item \code{terminators} :: \code{list()}\cr
List of objects of class \link{Terminator}.
\item \code{any} :: \code{logical(1)}\cr
Terminate iff any included terminator is positive? (not all).
Stored in \code{settings}.
}
}

\examples{
TerminatorCombo$new(list(
  TerminatorModelTime$new(60),
  TerminatorEvals$new(10)
))
}
\seealso{
Other Terminator: \code{\link{TerminatorClockTime}},
  \code{\link{TerminatorEvals}},
  \code{\link{TerminatorModelTime}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{Terminator}}, \code{\link{mlr_terminators}},
  \code{\link{mlr_tuners}}
}
\concept{Terminator}
\keyword{datasets}
