% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvaluations.R
\name{TerminatorEvaluations}
\alias{TerminatorEvaluations}
\title{TerminatorEvaluations}
\description{
TerminatorEvaluations.
}
\section{Usage}{
\preformatted{l = Terminator(id)
# public members
# public methods
# active bindings
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The name of the Terminator.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorEvaluations}.
}

\seealso{
Other Terminator: \code{\link{TerminatorBase}},
  \code{\link{TerminatorIterations}},
  \code{\link{TerminatorMultiplexer}}
}
\concept{Terminator}
\keyword{internal}
