% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvaluations.R
\docType{data}
\name{TerminatorEvaluations}
\alias{TerminatorEvaluations}
\title{Terminator that stops after a number of evals}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning depending on the number of evaluations.
An evaluation is defined by one resampling of a parameter value.
}
\section{Construction}{
\preformatted{t = TerminatorEvaluations$new(evals)
}
\itemize{
\item \code{evals} :: \code{integer(1)}\cr
Number of allowed evaluations.
}
}

\examples{
TerminatorEvaluations$new(3)
}
\seealso{
Other Terminator: \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
