% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorMultiplexer.R
\docType{data}
\name{TerminatorMultiplexer}
\alias{TerminatorMultiplexer}
\title{Combine Terminators}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
This class takes multiple \link{Terminator}s and terminates as soon as one of the subordinate Terminators signals a termination.
}
\section{Construction}{
\preformatted{t = TerminatorMultiplexer$new(terminators)
}
\itemize{
\item \code{terminators} :: \code{list()}\cr
List of objects of class \link{Terminator}.
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
t = TerminatorMultiplexer$new(list(
  TerminatorRuntime$new(60),
  TerminatorEvaluations$new(10)
))
print(t)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
