% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerfReached.R
\docType{data}
\name{TerminatorPerfReached}
\alias{TerminatorPerfReached}
\title{Terminator that stops when a peformance level has been reached}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a performance level has been hit.
}
\section{Construction}{
\preformatted{t = TerminatorPerfReached$new(level)
}
\itemize{
\item \code{level} :: \code{numeric(1)}\cr
Level that needs to be reached.
Terminates if the performance exceeds (respective measure has to be maximized) or
falls below (respective measure has to be minimized).
}
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorRuntime}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
