% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerfReached.R
\name{TerminatorPerfReached}
\alias{TerminatorPerfReached}
\alias{mlr_terminators_perf_reached}
\title{Terminator that stops when a performance level has been reached}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a performance level has been hit.
}
\section{Construction}{
\preformatted{TerminatorPerfReached$new()
term("perf_reached")
}
}

\section{Parameters}{

\itemize{
\item \code{level} :: \code{numeric(1)}\cr
Performance level that needs to be reached, default is 0.
Terminates if the performance exceeds (respective measure has to be maximized) or
falls below (respective measure has to be minimized) this value.
}
}

\examples{
TerminatorPerfReached$new()
term("perf_reached")
}
\seealso{
Other Terminator: 
\code{\link{TerminatorClockTime}},
\code{\link{TerminatorCombo}},
\code{\link{TerminatorEvals}},
\code{\link{TerminatorModelTime}},
\code{\link{TerminatorNone}},
\code{\link{TerminatorStagnation}},
\code{\link{Terminator}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
