% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerfReached.R
\docType{data}
\name{TerminatorPerfReached}
\alias{TerminatorPerfReached}
\alias{mlr_terminators_perf_reached}
\title{Terminator that stops when a performance level has been reached}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a performance level has been hit.
}
\section{Construction}{
\preformatted{TerminatorPerfReached$new(level)
term("perf_reached")
}
\itemize{
\item \code{measure} :: \code{character(1)}\cr
Name of a measure in the \link{TuningInstance}.
Stored in the parameter set \code{$param_set}.
\item \code{level} :: \code{numeric(1)}\cr
Performance level that needs to be reached.
Terminates if the performance exceeds (respective measure has to be maximized) or
falls below (respective measure has to be minimized).
Stored in the parameter set \code{$param_set}.
}
}

\seealso{
Other Terminator: \code{\link{TerminatorClockTime}},
  \code{\link{TerminatorCombo}},
  \code{\link{TerminatorEvals}},
  \code{\link{TerminatorModelTime}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorStagnation}},
  \code{\link{Terminator}}, \code{\link{mlr_terminators}},
  \code{\link{mlr_tuners}}
}
\concept{Terminator}
\keyword{datasets}
