% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorRuntime.R
\docType{data}
\name{TerminatorRuntime}
\alias{TerminatorRuntime}
\title{Terminator that stops afthat stops after runtime is depletedd}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a given runtime budget is exceeded.
}
\section{Construction}{
\preformatted{t = TerminatorRuntime$new(runtime)
}
\itemize{
\item \code{runtime} :: \code{numeric(1)}\cr
Maximum allowed runtime, in seconds.
}
}

\examples{
TerminatorRuntime$new(3)
TerminatorRuntime$new(10 * 3600)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorNone}},
  \code{\link{TerminatorPerfReached}},
  \code{\link{TerminatorTime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
