% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorRuntime.R
\docType{data}
\name{TerminatorRuntime}
\alias{TerminatorRuntime}
\title{Terminator with Runtime Criterion}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning after a given runtime budget is exceeded.
Runtime measurements starts with \code{update_start()} and ends with \code{update_end()}.
Note that the runtime is checked after each step and therefore it is possible that the final runtime is longer than the specified one.
}
\section{Construction}{
\preformatted{t = TerminatorRuntime$new(runtime)
}
\itemize{
\item \code{time} :: \code{numeric(1)}\cr
Maximum allowed runtime, in seconds.
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
t = TerminatorRuntime$new(3)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorMultiplexer}},
  \code{\link{TerminatorPerformance}},
  \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
