% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{tnr}
\alias{tnr}
\alias{term}
\title{Syntactic Sugar for Tuner and Terminator Construction}
\usage{
tnr(.key, ...)

term(.key, ...)
}
\arguments{
\item{.key}{:: \code{character(1)}\cr
Key passed to the respective \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} to retrieve the object.}

\item{...}{:: named \code{list()}\cr
Named arguments passed to the constructor, to be set as parameters in the \link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar]{mlr3misc::dictionary_sugar()}} for more details.}
}
\value{
\link{Tuner} for \code{tnr()} and \link{Terminator} for \code{term()}.
}
\description{
This function complements \link{mlr_tuners} and \link{mlr_terminators} with functions in the spirit
of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
term("evals", n_evals = 10)
tnr("random_search")
}
