% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{tune}
\alias{tune}
\title{Syntactic Sugar for Tuning}
\usage{
tune(
  method,
  task,
  learner,
  resampling,
  measures,
  term_evals = NULL,
  term_time = NULL,
  search_space = NULL,
  ...
)
}
\description{
Sugar
}
