% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuner.R
\docType{data}
\name{Tuner}
\alias{Tuner}
\title{Tuner}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Abstract \code{Tuner} class that implements the main functionality each tuner must have.
A tuner is an object that describes the tuning strategy how to search the hyperparameter space given within
the \code{[PerfEval]} object.
The state of tuning is stored in field \code{pe$bmr} and the tuner offers some active
bindings to conveniently access results.
}
\section{Construction}{
\preformatted{tuner = Tuner$new(pe, terminator, settings = list())
}
\itemize{
\item \code{pe} :: \link{PerfEval}.
\item \code{terminator} :: \link{Terminator}.
\item \code{settings} :: named \code{list()}\cr
Arbitrary list, depending on the child class.
}
}

\section{Fields}{

\itemize{
\item \code{pe} :: \link{PerfEval}\cr
The current state of the \link{PerfEval}.
\item \code{terminator} :: \link{Terminator}.
The current state of the \link{Terminator}.
\item \code{settings} :: named \code{list()}\cr
Arbitrary list, depending on the child class.
}
}

\section{Methods}{

\itemize{
\item \code{tune()}\cr
() -> \code{self}\cr
Performs the tuning until the \link{Terminator} becomes positive.
\item \code{tune_result(ties.method = "random")}\cr
(\code{character(1)}) -> named \code{list()}\cr
\code{ties_method} can be "first", "last" or "random".
List with 2 elements:
\itemize{
\item \code{performance} (\code{numeric()}) with the best performance.
\item \code{values} (named \code{list()}) with the corresponding hyperparameters values.
}
\item \code{archive(unnest = TRUE)}\cr
(\code{logical(1)}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a table of contained resample results, simply delegates to the \code{archive} method of \link{PerfEval}.
\item \code{eval_batch(dt)}\cr
(\code{\link[data.table:data.table]{data.table::data.table()}}) -> \link[data.table:data.table]{data.table::data.table}\cr
Evaluates a set of design points, see \code{eval_batch} in \link{PerfEval}, which this method delegates to.
This is the only entry point a subclass tuner should use to evaluate the objective function,
and it should normally not be called from the outside by the user.
The difference to \code{eval_batch} in \link{PerfEval} is that this also runs all associated terminators and potentially
stops with an exception.
}
}

\section{Technical Details and Subclasses}{

A subclass is implemented in the following way:
\itemize{
\item Inherit from Tuner
\item Specify the private abstract method \code{tune_internal} and use it to call into your optimizer.
\item When you set up an objective function, you will call \code{eval_batch} to evaluate design points.
\item The batch-eval is requested at the PerfEval 'pe' object,
so each batch is possibly executed in parallel via \code{\link[mlr3:benchmark]{mlr3::benchmark()}},
and all evaluations are stored inside of 'pe$bmr'.
\item After the batch-eval, all terminators registered with the tuner are checked, and if one is positive,
an exception is generated of class 'terminated_message'. In this case the current
batch of evals is still stored in pe, but the numeric score are not sent back to t
he handling optimizer as it has lost execution control.
\item After such an exception was caught we select the best configuration from \code{pe$bmr} and
return it.
\item Note that therefore more points than specified by the Terminator might be evaluated,
as the Terminator is only checked after a batch-eval. How many more depends on the batchsize.
}
}

\examples{
library(mlr3)
library(paradox)
param_set = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
pe = PerfEval$new("iris", "classif.rpart", "holdout", "classif.ce", param_set)
terminator = TerminatorEvaluations$new(3)
tt = TunerRandomSearch$new(pe, terminator) # swap this line to use a different Tuner
tt$tune()
tt$tune_result() # returns best configuration and performance
tt$archive() # allows access of data.table / benchmark result of full path of all evaluations
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}
}
\concept{Tuner}
\keyword{datasets}
