% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuner.R
\docType{data}
\name{Tuner}
\alias{Tuner}
\title{Tuner}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Abstract \code{Tuner} class that implements the main functionality each tuner must have.
A tuner is an object that describes the tuning strategy how to search the hyperparameter space given within
the \code{[PerfEval]} object.
The state of tuning is stored in field \code{pe$bmr} and the tuner offers some active
bindings to conveniently access results.
}
\section{Construction}{
\preformatted{tuner = Tuner$new(pe, terminator, settings = list())
}
\itemize{
\item \code{pe} :: \link{PerfEval}.
\item \code{terminator} :: \link{Terminator}.
\item \code{settings} :: named \code{list()}\cr
Arbitrary list, depending on the child class.
}
}

\section{Fields}{

\itemize{
\item \code{pe} :: \link{PerfEval}\cr
The current state of the \link{PerfEval}.
\item \code{terminator} :: \link{Terminator}.
The current state of the \link{Terminator}.
\item \code{settings} :: named \code{list()}\cr
Arbitrary list, depending on the child class.
}
}

\section{Methods}{

\itemize{
\item \code{tune()}\cr
() -> \code{self}\cr
Performs the tuning until the \link{Terminator} becomes positive.
\item \code{tune_result()}\cr
() -> named \code{list()}\cr
List with 2 elements:
\itemize{
\item \code{performance} (\code{numeric()}) with the best performance.
\item \code{values} (named \code{list()}) with the corresponding hyperparameters values.
}
\item \code{aggregate(unnest = TRUE)}
\code{logical(1)} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a table of contained resample results, simply delegates to the \code{aggregate} method of \link{PerfEval}.
\item \code{eval_batch(dt)}
`\code{\link[data.table:data.table]{data.table::data.table()}} --> numeric(n)\cr
Evaluates a set of design points, passed as a data.table of n rows and returns n performance values.
This is the only entry point a subclass tuner should use to evaluate the objective function,
and it should normally not be called from the outside by the user.
}
}

\examples{
library(mlr3)
library(paradox)
resampling = mlr_resamplings$get("cv", param_vals = list(folds = 2))
param_set = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
pe = PerfEval$new("iris", "classif.rpart", resampling, "classif.ce", param_set)
terminator = TerminatorEvaluations$new(3)
tt = TunerRandomSearch$new(pe, terminator) # swap this line to use a different Tuner
tt$tune()
tt$tune_result() # returns best configuration and performance
tt$aggregate() # allows access of data.table / benchmark result of full path of all evaluations
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}
}
\concept{Tuner}
\keyword{datasets}
