% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuner.R
\docType{data}
\name{Tuner}
\alias{Tuner}
\title{Tuner}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Abstract \code{Tuner} class that implements the main functionality each tuner must have.
A tuner is an object that describes the tuning strategy how to search the hyperparameter space given within
the \code{[PerformanceEvaluator]} object.
}
\section{Construction}{
\preformatted{tuner = Tuner$new(id, pe, terminator, settings = list())
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Name of the tuner.
\item \code{pe} :: \link{PerformanceEvaluator}.
\item \code{terminator} :: \link{Terminator}.
\item \code{settings} :: named \code{list()}\cr
Arbitrary list, depending on the child class.
}
}

\section{Fields}{

\itemize{
\item \code{id} :: \code{character(1)}\cr
Name of the tuner.
\item \code{pe} :: \link{PerformanceEvaluator}\cr
The current state of the \link{PerformanceEvaluator}.
\item \code{terminator} :: \link{Terminator}.
The current state of the \link{Terminator}.
\item \code{settings} :: named \code{list()}\cr
Arbitrary list, depending on the child class.
}
}

\section{Methods}{

\itemize{
\item \code{tune()}\cr
() -> \code{self}\cr
Performs the tuning until the budget of the \link{Terminator} is exhausted.
\item \code{tune_result()}\cr
() -> named \code{list()}\cr
List with 2 elements:
\itemize{
\item \code{performance} (\code{numeric()}) with the best performance.
\item \code{values} (named \code{list()} with the corresponding hyperparameters values.
}
\item \code{aggregate(unnest = TRUE)}
\code{logical(1)} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a table of resample results, similar to the one returned by \code{\link[mlr3:benchmark]{mlr3::benchmark()}}'s \code{aggregate()} method.
If \code{unnest} is \code{TRUE}, hyperparameter settings are stored in separate columns instead of inside a list column
}
}

\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}
}
\concept{Tuner}
\keyword{datasets}
