% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuner.R
\name{Tuner}
\alias{Tuner}
\title{Abstract Tuner Class}
\description{
Abstract \code{Tuner} class that implements the main functionality each tuner must have. A tuner is
an object that describes the tuning strategy how to search the hyperparameter space given within
the \code{[FitnessFunction]} object.
}
\section{Usage}{
\preformatted{# Construction
tuner = Tuner$new(id, ff, terminator, settings = list())

# public members
tuner$id
tuner$ff
tuner$terminator
tuner$settings

# public methods
tuner$tune()
tuner$tune_result()

# active bindings
tuner$aggregated
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):\cr
The id of the Tuner.
\item \code{ff} (\code{[FitnessFunction]}).
\item \code{terminator} (\code{[Terminator]}).
\item \code{settings} (\code{list}):\cr
The settings for the Tuner.
}
}

\section{Details}{

\itemize{
\item \code{$new()} creates a new object of class \code{[Tuner]}.
\item \code{id} stores an identifier for this \code{[Tuner]}.
\item \code{ff} stores the \link{FitnessFunction} to optimize.
\item \code{terminator} stores the \code{[Terminator]}.
\item \code{settings} is a list of hyperparamter settings for this \code{[Tuner]}.
\item \code{tune()} performs the tuning, until the budget of the \code{[Terminator]} in the \code{[FitnessFunction]} is exhausted.
\item \code{tune_result()} returns a list with 2 elements:
\itemize{
\item \code{performance} (\code{numeric()}) with the best performance.
\item \code{param_vals} (\code{numeric()}) with corresponding hyperparameters.
}
}
}

\seealso{
Other Tuner: \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}
}
\concept{Tuner}
