% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGenSA.R
\docType{data}
\name{TunerGenSA}
\alias{TunerGenSA}
\title{TunerGenSA}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Tuner child class to conduct generalized simulated annealing (GenSA) as tuning technique.
Internally relies on \code{\link[GenSA:GenSA]{GenSA::GenSA()}} from package \pkg{GenSA}.
}
\section{Construction}{
\preformatted{tuner = TunerGenSA$new(pe, terminator, ...)
}

For arguments, see \link{Tuner}, and additionally:
\itemize{
\item \code{...} :: named \code{list()}\cr
Settings passed down do \code{\link[GenSA:GenSA]{GenSA::GenSA()}}
Default settings are \code{smooth = FALSE}, \code{acceptance.param = -15},
\code{simple.function = FALSE}, and \code{temperature = 250}.
}
}

\section{Fields}{

See \link{Tuner}.
}

\section{Methods}{

See \link{Tuner}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv")
resampling$param_set$values$folds = 2
measures = mlr3::mlr_measures$mget("classif.ce")
param_set = paradox::ParamSet$new(
  params = list(
    paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1)
  )
)
pe = PerformanceEvaluator$new(task, learner, resampling, measures, param_set)

terminator = TerminatorEvaluations$new(10)
rs = TunerGenSA$new(pe, terminator)
rs$tune()$tune_result()
}
\seealso{
Other Tuner: \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
