% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGenSA.R
\name{TunerGenSA}
\alias{TunerGenSA}
\alias{mlr_tuners_gensa}
\title{TunerGenSA}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Subclass for generalized simulated annealing tuning calling \code{\link[GenSA:GenSA]{GenSA::GenSA()}} from package \CRANpkg{GenSA}.
}
\section{Construction}{
\preformatted{TunerGenSA$new()
tnr("gensa")
}
}

\section{Parameters}{

\itemize{
\item \code{smooth} :: \code{logical(1)}\cr
\item \code{temperature} :: \code{numeric(1)}\cr
\item \code{acceptance.param} :: \code{numeric(1)}\cr
\item \code{verbose} :: \code{logical(1)}\cr
\item \code{trace.mat} :: \code{logical(1)}\cr
}

For the meaning of the control parameters, see \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
Note that we have removed all control parameters which refer to the termination of the algorithm and
where our terminators allow to obtain the same behavior.
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: 
\code{\link{TunerDesignPoints}},
\code{\link{TunerGridSearch}},
\code{\link{TunerRandomSearch}},
\code{\link{Tuner}},
\code{\link{mlr_tuners}}
}
\concept{Tuner}
