% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGenSA.R
\docType{data}
\name{TunerGenSA}
\alias{TunerGenSA}
\alias{mlr_tuners_gensa}
\title{TunerGenSA}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Subclass for generalized simulated annealing tuning with \code{\link[GenSA:GenSA]{GenSA::GenSA()}}.
}
\section{Construction}{
\preformatted{TunerGenSA$new(...)
tnr("gensa")
}
}

\section{Parameters}{

\itemize{
\item \code{control}\cr
Settings passed down do \code{\link[GenSA:GenSA]{GenSA::GenSA()}}
Default settings are \code{smooth = FALSE}, \code{acceptance.param = -15},
\code{simple.function = FALSE}, and \code{temperature = 250}.
}
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: \code{\link{TunerGridSearch}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
