% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGridSearch.R
\docType{data}
\name{TunerGridSearch}
\alias{TunerGridSearch}
\title{TunerGridSearch}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Tuner child class to conduct a grid search.
}
\section{Construction}{
\preformatted{tuner = TunerGridSearch$new(pe, terminator, resolution)
}

For arguments, see \link{Tuner}, and additionally:
\itemize{
\item \code{resolution} :: \code{integer(1)}\cr
Resolution of the grid.
If none is specified we will try to calculate the resolution form the Terminator.
}
}

\section{Fields}{

See \link{Tuner}.
}

\section{Methods}{

See \link{Tuner}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv")
resampling$param_set$values$folds = 2
measures = mlr3::mlr_measures$mget("classif.ce")
param_set = paradox::ParamSet$new(
  params = list(
    paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1)
  )
)
pe = PerformanceEvaluator$new(task, learner, resampling, measures, param_set)

terminator = TerminatorEvaluations$new(10)
gs = TunerGridSearch$new(pe, terminator)
gs$tune()$tune_result()
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
