% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGridSearch.R
\docType{data}
\name{TunerGridSearch}
\alias{TunerGridSearch}
\title{TunerGridSearch}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Subclass for grid search tuning.

The grid is constructed as a Cartesian product over discretized values per parameter,
see \link[paradox:generate_design_grid]{paradox::generate_design_grid}.
The grid is searched in random order.

In order to support general termination criteria and parallelization,
we evaluate points in a batch-fashion of size \code{batch_size}.
Number of \code{batch_size} points are evaluated per iteration (potentially in parallel),
then the termination criteria are checked.
}
\section{Construction}{
\preformatted{tuner = TunerGridSearch$new(pe, terminator, resolution = 10L, batch_size = 1L)
}

For arguments, see \link{Tuner}, and additionally:
\itemize{
\item \code{resolution} :: \code{integer(1)}\cr
Resolution of the grid, see \link[paradox:generate_design_grid]{paradox::generate_design_grid}.
Stored in \code{settings}.
\item \code{batch_size} :: \code{integer(1)}\cr
Maximum number of configurations to try in a batch.
Stored in \code{settings}.
}
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerRandomSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
