% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerGridSearch.R
\name{TunerGridSearch}
\alias{TunerGridSearch}
\title{TunerGridSearch}
\description{
TunerGridSearch
}
\section{Usage}{
\preformatted{tuner = TunerGridSearch$new(ff, resolution)
}

See \link{Tuner} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The id of the Tuner.
\item \code{resolution} (\code{integer(1)}):
Resolution of the grid.
If none is specified we will try to calculate the resolution form the Terminator.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TunerGridSearch}.
The interface is described in \link{Tuner}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv")
measures = mlr3::mlr_measures$mget("mmce")
terminator = TerminatorEvaluations$new(10)
param_set = paradox::ParamSet$new(params = list(paradox::ParamReal$new("cp", lower = 0.001, upper = 0.1)))

ff = FitnessFunction$new(task, learner, resampling, measures, param_set, terminator)

rs = TunerGridSearch$new(ff)
rs$tune()$tune_result()
}
\seealso{
Other Tuner: \code{\link{TunerRandomSearch}},
  \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{internal}
