% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerRandomSearch.R
\name{TunerRandomSearch}
\alias{TunerRandomSearch}
\title{TunerRandomSearch}
\description{
TunerRandomSearch
}
\section{Usage}{
\preformatted{tuner = TunerRandomSearch(id)
# public members
# public methods
# active bindings
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The id of the Tuner.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TunerRandomSearch}.
}

\seealso{
Other Tuner: \code{\link{TunerBase}}
}
\concept{Tuner}
\keyword{internal}
