% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerRandomSearch.R
\name{TunerRandomSearch}
\alias{TunerRandomSearch}
\title{TunerRandomSearch}
\description{
TunerRandomSearch
}
\section{Usage}{
\preformatted{tuner = TunerRandomSearch(id)
# public members
# public methods
# active bindings
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
The id of the Tuner.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TunerRandomSearch}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv")
measures = mlr3::mlr_measures$mget("mmce")
terminator = TerminatorEvaluations$new("term-evals", 10)
param_set = paradox::ParamSet$new(params = list(paradox::ParamReal$new("cp", lower = 0.001, upper = 0.1)))

ff = FitnessFunction$new(task, learner, resampling, measures, param_set, terminator)

rs = TunerRandomSearch$new("rs", ff)
rs$tune(ps)
}
\seealso{
Other Tuner: \code{\link{TunerBase}}
}
\concept{Tuner}
\keyword{internal}
