% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerRandomSearch.R
\docType{data}
\name{TunerRandomSearch}
\alias{TunerRandomSearch}
\title{TunerRandomSearch}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Tuner child class to conduct a random search.
}
\section{Construction}{
\preformatted{tuner = TunerRandomSearch$new(pe, terminator, batch_size = 100L)
}

For arguments, see \link{Tuner}, and additionally:
\itemize{
\item \code{batch_size} :: \code{integer(1)}\cr
Maximum number of configurations to try in a batch.
}
}

\section{Fields}{

See \link{Tuner}.
}

\section{Methods}{

See \link{Tuner}.
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerGridSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
