% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerRandomSearch.R
\docType{data}
\name{TunerRandomSearch}
\alias{TunerRandomSearch}
\title{TunerRandomSearch}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Subclass for random search tuning.

In order to support general termination criteria and parallelization,
we evaluate points in a batch-fashion of size \code{batch_size}.
Number of \code{batch_size} points are evaluated per iteration (potentially in parallel),
then the termination criteria are checked.
}
\section{Construction}{
\preformatted{tuner = TunerRandomSearch$new(batch_size = 1L)
}
\itemize{
\item \code{batch_size} :: \code{integer(1)}\cr
Maximum number of configurations to try in a batch.
Stored in \code{settings}.
}
}

\examples{
# see ?Tuner
}
\seealso{
Other Tuner: \code{\link{TunerGenSA}},
  \code{\link{TunerGridSearch}}, \code{\link{Tuner}}
}
\concept{Tuner}
\keyword{datasets}
